sdm660-common: fix sensorservcie
W libc    : Unable to set property "ctl.interface_start" to "android.frameworks.sensorservice@1.0::ISensorManager/default": error code: 0x20
I hwservicemanager: Tried to start android.frameworks.sensorservice@1.0::ISensorManager/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.